+++ R/broom.helpers-package.R
@@ -93,3 +93,13 @@ utils::globalVariables("where")
     ) %>%
     tidy_attach_model(model)
 }
+
+# attributes to be saved between tidy_* functions
+.save_attributes <- function(x) {
+  .attributes <- attributes(x)
+  .attributes_names <- intersect(
+    names(.attributes),
+    c("exponentiate", "coefficients_type", "coefficients_label")
+  )
+  .attributes[.attributes_names]
+}