+++ R/tidy_add_coefficients_type.R
@@ -9,7 +9,6 @@
 #' coefficient estimates. It should be consistent with the original call to
 #' [broom::tidy()]
 #' @param model the corresponding model, if not attached to `x`
-#' @inheritParams tidy_plus_plus
 #' @export
 #' @family tidy_helpers
 #' @examples
@@ -38,6 +37,9 @@ tidy_add_coefficients_type <- function(
     stop("'model' is not provided. You need to pass it or to use 'tidy_and_attach()'.")
   }
 
+  .attributes <- .save_attributes(x)
+  .attributes$exponentiate <- exponentiate
+
   coefficients_type <- model_get_coefficients_type(model)
   if (exponentiate)
     coefficients_label <- dplyr::case_when(
@@ -58,6 +60,5 @@ tidy_add_coefficients_type <- function(
   attr(x, "coefficients_label") <- coefficients_label
 
   x %>%
-    tidy_attach_model(model = model) %>%
-    .order_tidy_columns()
+    tidy_attach_model(model = model, .attributes = .attributes)
 }